Turn a jagam model definition into a greta model definition

Takes a GAM defined by formula and returns the corresponding greta model via the power of jagam. Response variable is generated from dummy data and not used.

Usage

jagam2greta(formula, data, newdata, sp = NULL, knots = NULL, tol = 0)

Arguments

formula

a GAM formula representing the smooth terms, as in mgcv::gam(). Only the right hand side fo the formula will be used.

data

a data frame or list containing the covariates required by the formula. These covariates cannot be greta arrays.

newdata

new dataset

sp

an optional vector of smoothing parameters, two per smooth term in the model, in the same order as the formula. If sp = NULL, all smoothing parameters will be learned, otherwise all smoothing parameters must be specified by the user. The smoothing parameters may either be a numeric vector or a greta array (which could be a variable).

knots

an optional list containing user specified knot values to be used for basis construction, as in mgcv::gam() These knots cannot be greta arrays.

tol

a non-negative scalar numerical tolerance parameter. You can try increasing this if the model has numerical stability issues

Value

a list with the following elements: betas a greta array for the coefficients to be estimated (with appropriate priors applied), X design matrix for this model, X_pred prediction matrix.